move box and plain distance to one file.
