+++ scipy/spatial/ckdtree/src/distance.h
@@ -1,4 +1,6 @@
-struct Dist1D {
+#include "distance_base.h"
+
+struct PlainDist1D {
     static inline void
     interval_interval(const ckdtree * tree,
                         const Rectangle& rect1, const Rectangle& rect2,
@@ -22,228 +24,10 @@ struct Dist1D {
     }
 };
 
-template <typename Dist1D>
-struct BaseMinkowskiDistP1 {
-
-    static inline void
-    interval_interval_p(const ckdtree * tree,
-                        const Rectangle& rect1, const Rectangle& rect2,
-                        const npy_intp k, const npy_float64 p,
-                        npy_float64 *min, npy_float64 *max)
-    {
-        /* Compute the minimum/maximum distance along dimension k between points in
-         * two hyperrectangles.
-         */
-        Dist1D::interval_interval(tree, rect1, rect2, k, min, max);
-    }
-
-    static inline void
-    rect_rect_p(const ckdtree * tree,
-                        const Rectangle& rect1, const Rectangle& rect2,
-                        const npy_float64 p,
-                        npy_float64 *min, npy_float64 *max)
-    {
-        *min = 0.;
-        *max = 0.;
-        for(npy_intp i=0; i<rect1.m; ++i) {
-            npy_float64 min_, max_;
-
-            Dist1D::interval_interval(tree, rect1, rect2, i, &min_, &max_);
-
-            *min += min_;
-            *max += max_;
-        }
-    }
-
-    static inline npy_float64
-    distance_p(const ckdtree * tree,
-               const npy_float64 *x, const npy_float64 *y,
-               const npy_float64 p, const npy_intp k,
-               const npy_float64 upperbound)
-    {
-        npy_intp i;
-        npy_float64 r;
-        r = 0;
-        for (i=0; i<k; ++i) {
-            r += Dist1D::point_point(tree, x, y, i);
-            if (r>upperbound)
-                return r;
-        }
-        return r;
-    }
-};
-
-template <typename Dist1D>
-struct BaseMinkowskiDistPp {
-    /* 1-d pieces
-     * These should only be used if p != infinity
-     */
-
-    static inline void
-    interval_interval_p(const ckdtree * tree,
-                        const Rectangle& rect1, const Rectangle& rect2,
-                        const npy_intp k, const npy_float64 p,
-                        npy_float64 *min, npy_float64 *max)
-    {
-        /* Compute the minimum/maximum distance along dimension k between points in
-         * two hyperrectangles.
-         */
-        Dist1D::interval_interval(tree, rect1, rect2, k, min, max);
-        *min = std::pow(*min, p);
-        *max = std::pow(*max, p);
-    }
-
-    static inline void
-    rect_rect_p(const ckdtree * tree,
-                        const Rectangle& rect1, const Rectangle& rect2,
-                        const npy_float64 p,
-                        npy_float64 *min, npy_float64 *max)
-    {
-        *min = 0.;
-        *max = 0.;
-        for(npy_intp i=0; i<rect1.m; ++i) {
-            npy_float64 min_, max_;
-
-            Dist1D::interval_interval(tree, rect1, rect2, i, &min_, &max_);
-
-            *min += std::pow(min_, p);
-            *max += std::pow(max_, p);
-        }
-    }
-
-    static inline npy_float64
-    distance_p(const ckdtree * tree,
-               const npy_float64 *x, const npy_float64 *y,
-               const npy_float64 p, const npy_intp k,
-               const npy_float64 upperbound)
-    {
-       /*
-        * Compute the distance between x and y
-        *
-        * Computes the Minkowski p-distance to the power p between two points.
-        * If the distance**p is larger than upperbound, then any number larger
-        * than upperbound may be returned (the calculation is truncated).
-        */
-
-        npy_intp i;
-        npy_float64 r, r1;
-        r = 0;
-        for (i=0; i<k; ++i) {
-            r1 = Dist1D::point_point(tree, x, y, i);
-            r += std::pow(r1, p);
-            if (r>upperbound)
-                 return r;
-        }
-        return r;
-    }
-};
-
-template <typename Dist1D>
-struct BaseMinkowskiDistPinf {
-    static inline void
-    interval_interval_p(const ckdtree * tree,
-                        const Rectangle& rect1, const Rectangle& rect2,
-                        const npy_intp k, npy_float64 p,
-                        npy_float64 *min, npy_float64 *max)
-    {
-        return rect_rect_p(tree, rect1, rect2, p, min, max);
-    }
-
-    static inline void
-    rect_rect_p(const ckdtree * tree,
-                        const Rectangle& rect1, const Rectangle& rect2,
-                        const npy_float64 p,
-                        npy_float64 *min, npy_float64 *max)
-    {
-        *min = 0.;
-        *max = 0.;
-        for(npy_intp i=0; i<rect1.m; ++i) {
-            npy_float64 min_, max_;
-
-            Dist1D::interval_interval(tree, rect1, rect2, i, &min_, &max_);
-
-            *min = dmax(*min, min_);
-            *max = dmax(*max, max_);
-        }
-    }
-
-    static inline npy_float64
-    distance_p(const ckdtree * tree,
-               const npy_float64 *x, const npy_float64 *y,
-               const npy_float64 p, const npy_intp k,
-               const npy_float64 upperbound)
-    {
-        npy_intp i;
-        npy_float64 r;
-        r = 0;
-        for (i=0; i<k; ++i) {
-            r = dmax(r,Dist1D::point_point(tree, x, y, i));
-            if (r>upperbound)
-                return r;
-        }
-        return r;
-    }
-};
-
-template <typename Dist1D>
-struct BaseMinkowskiDistP2 {
-    static inline void
-    interval_interval_p(const ckdtree * tree,
-                        const Rectangle& rect1, const Rectangle& rect2,
-                        const npy_intp k, const npy_float64 p,
-                        npy_float64 *min, npy_float64 *max)
-    {
-        /* Compute the minimum/maximum distance along dimension k between points in
-         * two hyperrectangles.
-         */
-        Dist1D::interval_interval(tree, rect1, rect2, k, min, max);
-        *min *= *min;
-        *max *= *max;
-    }
-
-    static inline void
-    rect_rect_p(const ckdtree * tree,
-                        const Rectangle& rect1, const Rectangle& rect2,
-                        const npy_float64 p,
-                        npy_float64 *min, npy_float64 *max)
-    {
-        *min = 0.;
-        *max = 0.;
-        for(npy_intp i=0; i<rect1.m; ++i) {
-            npy_float64 min_, max_;
-
-            Dist1D::interval_interval(tree, rect1, rect2, i, &min_, &max_);
-            min_ *= min_;
-            max_ *= max_;
-
-            *min += min_;
-            *max += max_;
-        }
-    }
-    static inline npy_float64
-
-    distance_p(const ckdtree * tree,
-               const npy_float64 *x, const npy_float64 *y,
-               const npy_float64 p, const npy_intp k,
-               const npy_float64 upperbound)
-    {
-        npy_intp i;
-        npy_float64 r;
-        r = 0;
-        for (i=0; i<k; ++i) {
-            npy_float64 r1 = Dist1D::point_point(tree, x, y, i);
-            r += r1 * r1;
-            if (r>upperbound)
-                return r;
-        }
-        return r;
-    }
-};
-
-typedef BaseMinkowskiDistPp<Dist1D> MinkowskiDistPp;
-typedef BaseMinkowskiDistPinf<Dist1D> MinkowskiDistPinf;
-typedef BaseMinkowskiDistP1<Dist1D> MinkowskiDistP1;
-typedef BaseMinkowskiDistP2<Dist1D> NonOptimizedMinkowskiDistP2;
+typedef BaseMinkowskiDistPp<PlainDist1D> MinkowskiDistPp;
+typedef BaseMinkowskiDistPinf<PlainDist1D> MinkowskiDistPinf;
+typedef BaseMinkowskiDistP1<PlainDist1D> MinkowskiDistP1;
+typedef BaseMinkowskiDistP2<PlainDist1D> NonOptimizedMinkowskiDistP2;
 
 struct MinkowskiDistP2: NonOptimizedMinkowskiDistP2 {
     static inline npy_float64
@@ -255,3 +39,114 @@ struct MinkowskiDistP2: NonOptimizedMinkowskiDistP2 {
         return sqeuclidean_distance_double(x, y, k);
     }
 };
+
+struct BoxDist1D {
+    static inline void _interval_interval_1d (
+        npy_float64 min, npy_float64 max,
+        npy_float64 *realmin, npy_float64 *realmax,
+        const npy_float64 full, const npy_float64 half
+    )
+    {
+        /* Minimum and maximum distance of two intervals in a periodic box
+         *
+         * min and max is the nonperiodic distance between the near
+         * and far edges.
+         *
+         * full and half are the box size and 0.5 * box size.
+         *
+         * value is returned in realmin and realmax;
+         *
+         * This function is copied from kdcount, and the convention
+         * of is that
+         *
+         * min = rect1.min - rect2.max
+         * max = rect1.max - rect2.min = - (rect2.min - rect1.max)
+         *
+         * We will fix the convention later.
+         * */
+        if (NPY_UNLIKELY(full <= 0)) {
+            /* A non-periodic dimension */
+            /* \/     */
+            if(max <= 0 || min >= 0) {
+                /* do not pass though 0 */
+                min = dabs(min);
+                max = dabs(max);
+                if(min < max) {
+                    *realmin = min;
+                    *realmax = max;
+                } else {
+                    *realmin = max;
+                    *realmax = min;
+                }
+            } else {
+                min = dabs(min);
+                max = dabs(max);
+                *realmax = fmax(max, min);
+                *realmin = 0;
+            }
+            /* done with non-periodic dimension */
+            return;
+        }
+        if(max <= 0 || min >= 0) {
+            /* do not pass through 0 */
+            min = dabs(min);
+            max = dabs(max);
+            if(min > max) {
+                double t = min;
+                min = max;
+                max = t;
+            }
+            if(max < half) {
+                /* all below half*/
+                *realmin = min;
+                *realmax = max;
+            } else if(min > half) {
+                /* all above half */
+                *realmax = full - min;
+                *realmin = full - max;
+            } else {
+                /* min below, max above */
+                *realmax = half;
+                *realmin = dmin(min, full - max);
+            }
+        } else {
+            /* pass though 0 */
+            min = -min;
+            if(min > max) max = min;
+            if(max > half) max = half;
+            *realmax = max;
+            *realmin = 0;
+        }
+    }
+    static inline void
+    interval_interval(const ckdtree * tree,
+                        const Rectangle& rect1, const Rectangle& rect2,
+                        const npy_intp k,
+                        npy_float64 *min, npy_float64 *max)
+    {
+        /* Compute the minimum/maximum distance along dimension k between points in
+         * two hyperrectangles.
+         */
+        _interval_interval_1d(rect1.mins()[k] - rect2.maxes()[k],
+                    rect1.maxes()[k] - rect2.mins()[k], min, max,
+                    tree->raw_boxsize_data[k], tree->raw_boxsize_data[k + rect1.m]);
+    }
+
+    static inline npy_float64
+    point_point(const ckdtree * tree,
+               const npy_float64 *x, const npy_float64 *y,
+               const npy_intp k)
+    {
+        npy_float64 r1;
+        r1 = wrap_distance(x[k] - y[k], tree->raw_boxsize_data[k + tree->m], tree->raw_boxsize_data[k]);
+        r1 = dabs(r1);
+        return r1;
+    }
+};
+
+
+typedef BaseMinkowskiDistPp<BoxDist1D> BoxMinkowskiDistPp;
+typedef BaseMinkowskiDistPinf<BoxDist1D> BoxMinkowskiDistPinf;
+typedef BaseMinkowskiDistP1<BoxDist1D> BoxMinkowskiDistP1;
+typedef BaseMinkowskiDistP2<BoxDist1D> BoxMinkowskiDistP2;
+
